more memory hoops

+++ train.py
@@ -11,5 +11,5 @@ train_dataloader = get_dataloader("domain_set")
 accelerator = "cuda" if torch.cuda.is_available() else "cpu"
 
 trainer = L.Trainer(max_epochs=config.num_epochs,
-                    accelerator=accelerator, accumulate_grad_batches=2)
+                    accelerator=accelerator, precision='16-mixed', accumulate_grad_batches=2) 
 trainer.fit(model, train_dataloader)
